refactor: move TooltipData type near to Tooltip component
