+++ src/chart_types/xy_chart/state/selectors/is_tooltip_visible.ts
@@ -3,11 +3,12 @@ import { Point } from '../../../../utils/point';
 import { GlobalChartState, PointerStates } from '../../../../state/chart_state';
 import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
 import { getProjectedPointerPositionSelector } from './get_projected_pointer_position';
-import { getTooltipValuesSelector, TooltipData } from './get_tooltip_values_highlighted_geoms';
+import { getTooltipValuesSelector } from './get_tooltip_values_highlighted_geoms';
 import { getChartIdSelector } from '../../../../state/selectors/get_chart_id';
 import { getTooltipType } from './get_tooltip_type';
 import { TooltipType } from '../../../../specs';
 import { isAnnotationTooltipVisibleSelector } from './is_annotation_tooltip_visible';
+import { TooltipData } from '../../../../components/tooltip/types';
 
 const hasTooltipTypeDefinedSelector = (state: GlobalChartState): TooltipType | undefined => {
   return getTooltipType(getSettingsSpecSelector(state));
