+++ src/components/tooltip/index.tsx
@@ -8,11 +8,9 @@ import { isInitialized } from '../../state/selectors/is_initialized';
 import { getTooltipHeaderFormatterSelector } from '../../state/selectors/get_tooltip_header_formatter';
 import { isTooltipVisibleSelector } from '../../chart_types/xy_chart/state/selectors/is_tooltip_visible';
 import { getTooltipPositionSelector } from '../../chart_types/xy_chart/state/selectors/get_tooltip_position';
-import {
-  getTooltipValuesSelector,
-  TooltipData,
-} from '../../chart_types/xy_chart/state/selectors/get_tooltip_values_highlighted_geoms';
+import { getTooltipValuesSelector } from '../../chart_types/xy_chart/state/selectors/get_tooltip_values_highlighted_geoms';
 import { getFinalTooltipPosition, TooltipPosition } from '../../chart_types/xy_chart/crosshair/crosshair_utils';
+import { TooltipData } from './types';
 
 interface TooltipStateProps {
   isTooltipVisible: boolean;
