+++ src/components/tooltip/types.ts
@@ -1,0 +1,6 @@
+import { TooltipValue } from '../../specs';
+
+export interface TooltipData {
+  header: TooltipValue | null;
+  values: TooltipValue[];
+}
